tf1x/requirements.txt
DeepSparseCoding
h5py>=2.8.0
imageio>=2.4.1
ipython>=7.1.1
jupyter>=1.0.0
matplotlib>=3.1.1
numpy>=1.15.3
pandas>=0.23.4
Pillow>=5.3.0
scikit-image>=0.14.1
scikit-learn>=0.20.0
scipy>=1.1.0
seaborn>=0.9.0
tensorboard==1.12.2
tensorflow-gpu==1.15.2
tensorflow-compression
